Multiple modifications
1. RNAseq pipeline style fix
2. Dataset construction: added TMM and QN normalization
3. Differential expression: added time series analyzer

+++ 01a_reads_dumping/launch.sh
@@ -29,7 +29,7 @@ input="rnaseq"
 
 # copia los ficheros de datos en $tempdir 
 cp -r "$submitdir"/$input "$tempdir"
-cd "$tempdir"/"$input"
+cd "$tempdir"/"$input" || exit
 # ejecuta tus comandos
 
 conda activate rnaseq
@@ -44,8 +44,8 @@ conda deactivate
 cp -r "$SCRDIR"/"$input" "$submitdir"
  
 # limpieza 
-cd $submitdir
-rm -r "$tempdir"/*
+cd "${submitdir}" || exit
+rm -r "${tempdir:?}"/*
 rmdir  "$tempdir"
  
 echo "Job finished at"
